--- tslib/src/functions/school-pin-generator/school-pin-generator.service.ts
@@ -1,32 +1,0 @@
-
-export class SchoolPinGenerator {
-
-  private dataService: ISchoolPinGeneratorDataService
-
-  constructor(dataService?: ISchoolPinGeneratorDataService) {
-    if (dataService === undefined) {
-      dataService = new SchoolPinGeneratorDataService()
-    }
-    this.dataService = dataService
-  }
-}
-
-export class SchoolPinGeneratorDataService implements ISchoolPinGeneratorDataService {
-  getSchools (): School[] {
-    throw new Error('Method not implemented.')
-  }
-
-  setPins (schools: Array<School>) {
-    throw new Error('not implemented')
-  }
-}
-
-export interface ISchoolPinGeneratorDataService {
-  getSchools (): Array<School>
-}
-
-export interface School {
-  id: number
-  name: string
-  utcOffset: number
-}
